cards/serializers.py
from rest_framework import serializers
from .models import Card


class CardSerializer(serializers.ModelSerializer):
    class Meta:
        model = Card
        fields = ('user', 'title_native_language', 'translate_studied_language', 'category')
