requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.14.0
Pillow==9.2.0
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.2.1
sqlparse==0.4.2
text-unidecode==1.3
